COMMERCE-11740 adapt usage

+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/method/DefaultCommerceInventoryMethodImpl.java
@@ -36,6 +36,8 @@ import com.liferay.portal.kernel.transaction.Propagation;
 import com.liferay.portal.kernel.transaction.Transactional;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
+import java.math.BigDecimal;
+
 import java.util.Locale;
 import java.util.Map;
 import java.util.ResourceBundle;
@@ -79,8 +81,8 @@ public class DefaultCommerceInventoryMethodImpl
 
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
-			commerceInventoryAuditType.getLog(context), quantity, sku,
-			StringPool.BLANK);
+			commerceInventoryAuditType.getLog(context),
+			BigDecimal.valueOf(quantity), sku, StringPool.BLANK);
 
 		for (CommerceInventoryEngineContributor
 				commerceInventoryEngineContributor :
@@ -220,8 +222,8 @@ public class DefaultCommerceInventoryMethodImpl
 
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
-			commerceInventoryAuditType.getLog(null), quantity, sku,
-			StringPool.BLANK);
+			commerceInventoryAuditType.getLog(null),
+			BigDecimal.valueOf(quantity), sku, StringPool.BLANK);
 
 		for (CommerceInventoryEngineContributor
 				commerceInventoryEngineContributor :
