+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/service/impl/CommerceInventoryBookedQuantityLocalServiceImpl.java
@@ -49,6 +49,8 @@ import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 
+import java.math.BigDecimal;
+
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -98,8 +100,8 @@ public class CommerceInventoryBookedQuantityLocalServiceImpl
 
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
-			commerceInventoryAuditType.getLog(context), quantity, sku,
-			StringPool.BLANK);
+			commerceInventoryAuditType.getLog(context),
+			BigDecimal.valueOf(quantity), sku, StringPool.BLANK);
 
 		return commerceInventoryBookedQuantityPersistence.update(
 			commerceInventoryBookedQuantity);
@@ -287,8 +289,8 @@ public class CommerceInventoryBookedQuantityLocalServiceImpl
 
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
-			commerceInventoryAuditType.getLog(context), quantity, sku,
-			StringPool.BLANK);
+			commerceInventoryAuditType.getLog(context),
+			BigDecimal.valueOf(quantity), sku, StringPool.BLANK);
 
 		return commerceInventoryBookedQuantityPersistence.update(
 			commerceBookedQuantity);
@@ -312,7 +314,7 @@ public class CommerceInventoryBookedQuantityLocalServiceImpl
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
 			commerceInventoryAuditType.getLog(context),
-			commerceInventoryBookedQuantity.getQuantity(),
+			BigDecimal.valueOf(commerceInventoryBookedQuantity.getQuantity()),
 			commerceInventoryBookedQuantity.getSku(), StringPool.BLANK);
 
 		return commerceInventoryBookedQuantity;
@@ -380,7 +382,8 @@ public class CommerceInventoryBookedQuantityLocalServiceImpl
 
 		_commerceInventoryAuditLocalService.addCommerceInventoryAudit(
 			userId, commerceInventoryAuditType.getType(),
-			commerceInventoryAuditType.getLog(context), quantity,
+			commerceInventoryAuditType.getLog(context),
+			BigDecimal.valueOf(quantity),
 			commerceInventoryBookedQuantity.getSku(),
 			commerceInventoryBookedQuantity.getUnitOfMeasureKey());
